* make PL_put_integer compatible with docs again

+++ src/pl-itf.h
@@ -407,7 +407,7 @@ PL_EXPORT(void)		PL_put_atom_nchars(term_t t, size_t l, const char *chars);
 PL_EXPORT(void)		PL_put_string_nchars(term_t t, size_t len, const char *chars);
 PL_EXPORT(void)		PL_put_list_nchars(term_t t, size_t l, const char *chars);
 PL_EXPORT(void)		PL_put_list_ncodes(term_t t, size_t l, const char *chars);
-PL_EXPORT(void)		PL_put_integer(term_t t, intptr_t i);
+PL_EXPORT(void)		PL_put_integer(term_t t, long i);
 PL_EXPORT(void)		PL_put_pointer(term_t t, void *ptr);
 PL_EXPORT(void)		PL_put_float(term_t t, double f);
 PL_EXPORT(void)		PL_put_functor(term_t t, functor_t functor);
